Move migrations to own cmd
