+++ Router/Slack_Events.go
@@ -1,4 +1,4 @@
-package endpoints
+package Router
 
 import (
 	"encoding/json"
