+++ cmd/server/main.go
@@ -4,29 +4,15 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/hex337/alex-koin-go/endpoints"
-	"github.com/hex337/alex-koin-go/Models"
 	"github.com/hex337/alex-koin-go/Config"
+	"github.com/hex337/alex-koin-go/Router"
 )
 
 func main() {
-	var err error
-
 	Config.GetEnvVars()
 	Config.DBOpen()
 
-	// defer config.DB.Close()
-	Config.DB.AutoMigrate(&Models.Transaction{}, &Models.User{}, &Models.Coin{})
-
-	var user []Models.User
-	err = Models.GetAllUsers(&user)
-	if err != nil {
-		log.Println(err.Error())
-	}
-
-	log.Printf("%v", user[0])
-
-	endpoints.SlackEvents()
+	Router.SlackEvents()
 
 	serverURL := Config.ServerURL(Config.BuildServerConfig())
 	log.Printf("[INFO] Server listening %s", serverURL)
